Improve Dependency Injection and Improve API Key Input
Dependency Injection is replaced with a model of
protocol composition following an article written
by Krzystztof Zablocki (@merowing_).

API Key input textField was replaced by a
UIAlertController popup. This seems to follow
the general input of "secure" data around the OS.

Created Alertable protocol which provides the ability
to show alerts in an easier way for UIViewControllers.